# cardgameserver/swagger/swagger.yaml
swagger: '2.0'
info:
  description: APIforcardGame.
  version: 0.0.1
  title: Swagger Card
  contact:
    email: lenon461@gmail.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
host: 'localhost:11220'
basePath: /
tags:
  - name: card
    description: test
    externalDocs:
      description: Find out more
      url: 'http://swagger.io'
schemes:
  - http
paths:
  /start:
    get:
      tags:
        - card
      summary: initialize card
      description: ''
      operationId: test
      consumes:
        - application/xml
      produces:
        - application/xml
      parameters:
        - in: body
          name: start a card
      responses:
        '200':
          description: success
  /divide6:
    put:
      tags:
        - card
      summary: initialize card
      description: ''
      operationId: test2
      consumes:
        - application/json
      produces:
        - application/xml
      parameters:
        - in: body
          name: divide 6 card
          schema:
              $ref: '#/definitions/Card'
      responses:
        '200':
          description: success
  /divide8:
    put:
      tags:
        - card
      summary: initialize card
      description: ''
      operationId: test2
      consumes:
        - application/json
      produces:
        - application/xml
      parameters:
        - in: body
          name: divide 8 card
          schema:
              $ref: '#/definitions/Card'
      responses:
        '200':
          description: success
  /raise:
    put:
      tags:
        - card
      summary: initialize card
      description: ''
      operationId: test3
      consumes:
        - application/json
      produces:
        - application/xml
      parameters:
        - in: query
          name: player
        - in: body
          name: body
          schema:
              $ref: '#/definitions/Card'
      responses:
        '200':
          description: success
  /acquire:
    put:
      tags:
        - card
      summary: initialize card
      description: ''
      operationId: test3
      consumes:
        - application/json
      produces:
        - application/xml
      parameters:
        - in: body
          name: body
          schema:
              $ref: '#/definitions/Card'
      responses:
        '200':
          description: success
securityDefinitions:
  petstore_auth:
    type: oauth2
    authorizationUrl: 'http://petstore.swagger.io/oauth/dialog'
    flow: implicit
    scopes:
      'write:pets': modify pets in your account
      'read:pets': read your pets
  api_key:
    type: apiKey
    name: api_key
    in: header
definitions:
  Card:
    type: object
    properties:
      peoples_hand:
        type: array
        items:
          type: array
          items:
            type: string
      peoples_acquired:
        type: array
        items:
          type: array
          items:
              type: string
      peoples_large_t:
        type: array
        items:
          type: integer
      peoples_small_t:
        type: array
        items:
          type: integer
      peoples_name:
        type: array
        items:
          type: string
      team1_member:
        type: array
        items:
          type: string
      team2_member:
        type: array
        items:
          type: string
      team1_point:
        type: integer
        default: 0
      team2_point:
        type: integer
        default: 0
      card_dummy:
        type: array
        items:
          type: string
      field_card:
        type: array
        items:
          type: string
      field_card_point:
        type: integer
        default: 0
      top_card:
        type: array
        items:
          type: string
      top_card_power:
        type: integer
        default: 0
      first_player:
        type: integer
        default: 0
      next_player:
        type: integer
        default: 0
externalDocs:
  description: Find out more about Swagger
  url: 'http://swagger.io'